backend/requirements.txt
Django>=1.11,<2.0
Pillow
psycopg2
djangorestframework
djangorestframework-jwt
django-cors-headers
